+++ core/healthcheck/healthcheck.go
@@ -1,0 +1,26 @@
+package healthcheck
+
+import (
+	"context"
+
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+
+	healthpb "google.golang.org/grpc/health/grpc_health_v1"
+)
+
+type HealthServer struct {
+	healthpb.UnimplementedHealthServer
+}
+
+func NewHealthServer() *HealthServer {
+	return &HealthServer{}
+}
+
+func (s *HealthServer) Check(_ context.Context, _ *healthpb.HealthCheckRequest) (*healthpb.HealthCheckResponse, error) {
+	return &healthpb.HealthCheckResponse{Status: healthpb.HealthCheckResponse_SERVING}, nil
+}
+
+func (s *HealthServer) Watch(_ *healthpb.HealthCheckRequest, _ healthpb.Health_WatchServer) error {
+	return status.Error(codes.Unimplemented, "Watching is not supported")
+}